Move docker module into `lib`
