+++ tests/test_docker_stats.py
@@ -2,7 +2,7 @@ import time
 
 import pytest
 
-from jobrunner import docker
+from jobrunner.lib import docker
 from jobrunner.docker_stats import get_container_stats, get_volume_and_container_sizes
