+++ tests/test_integration.py
@@ -6,7 +6,8 @@ import pytest
 
 import jobrunner.run
 import jobrunner.sync
-from jobrunner import config, docker
+from jobrunner import config
+from jobrunner.lib import docker
 from jobrunner.subprocess_utils import subprocess_run
 
 log = logging.getLogger(__name__)
